requirements.txt
Django
Pillow
jsonschema
django-ratelimit
django-widget-tweaks
pandas
openpyxl
defusedxml